+++ app-web.php
@@ -1,10 +1,18 @@
 <?php
 /**
  * @author Semenov Alexander <semenov@skeeks.com>
- * @link http://skeeks.com/
- * @copyright 2010 SkeekS (СкикС)
- * @date 31.03.2015
- * @var \skeeks\cms\Config $config
+ * @link https://skeeks.com/
+ * @copyright (c) 2010 SkeekS
+ * @date 10.11.2017
  */
-$application = require(__DIR__ . '/app-web-create.php');
+//Standard loader
+require(__DIR__ . '/bootstrap.php');
+
+$configFile = \hiqdev\composer\config\Builder::path('web-' . YII_ENV);
+if (!file_exists($configFile)) {
+    $configFile = \hiqdev\composer\config\Builder::path('web');
+}
+$config = (array) require $configFile;
+
+$application = new yii\web\Application($config);
 $application->run();