+++ bootstrap.php
@@ -1,14 +1,14 @@
 <?php
 /**
  * @author Semenov Alexander <semenov@skeeks.com>
- * @link http://skeeks.com/
- * @copyright 2010 SkeekS (СкикС)
- * @date 17.04.2016
+ * @link https://skeeks.com/
+ * @copyright (c) 2010 SkeekS
+ * @date 10.11.2017
  */
+require(__DIR__ . '/global.php');
+
 require(VENDOR_DIR . '/autoload.php');
 require(VENDOR_DIR . '/yiisoft/yii2/Yii.php');
-require(COMMON_CONFIG_DIR . '/bootstrap.php');
-require(APP_CONFIG_DIR . '/bootstrap.php');
 
-\Yii::setAlias('@root', ROOT_DIR);
-\Yii::setAlias('@vendor', VENDOR_DIR);
+\Yii::setAlias('@root',     ROOT_DIR);
+\Yii::setAlias('@vendor',   VENDOR_DIR);